requirements.txt
pandas
pypdf
pymongo>=4.0
python-docx
python-pptx
google-generativeai
python-dotenv
flask
werkzeug
pillow
openai>=1.0.0
